--- base/TorqueState.java
@@ -1,8 +1,0 @@
-package org.texastorque.torquelib.base;
-
-/**
- * A enum interface that is extended by enums that 
- * describe subsystem state.
- */
-public interface TorqueState {
-}
