+++ base/TorqueSubsystem.java
@@ -8,9 +8,6 @@ package org.texastorque.torquelib.base;
  */
 public abstract class TorqueSubsystem {
 
-    public interface TorqueSubsystemState {
-    }
-
     public abstract void initTeleop();
     public abstract void updateTeleop();
